feat(rts-core): migrate `sync` to #[rts_namespace]
Stage 2c: collapse the sync namespace quad-write (abi.rs MEMBERS+SPEC, mutex/rwlock/once extern fns, codegen SPECS array, jit add_fn!) into a single annotated impl SyncNs in mod.rs.

- sync/mod.rs: #[rts_namespace(sync)] with 13 #[rts_fn] members (Mutex/RwLock/OnceLock)
- thread-local guard maps + Arc-anchoring soundness (#280) kept at module scope
- delete sync/abi.rs + sync/mutex.rs + sync/rwlock.rs + sync/once.rs
- rts-codegen abi/mod.rs: sync::abi::SPEC -> sync::SPEC
- rts-codegen jit.rs: collapse submodule uses into use sync

Suite 1719/1719 green, rts.d.ts unchanged. Refs #1378.
