--- crates/rts-runtime/src/namespaces/sync/abi.rs
@@ -1,147 +1,0 @@
-//! `sync` namespace — ABI registration.
-
-use crate::abi::{AbiType, MemberKind, NamespaceMember, NamespaceSpec};
-
-pub const MEMBERS: &[NamespaceMember] = &[
-    // ── Mutex ──────────────────────────────────────────────────────
-    NamespaceMember {
-        name: "mutex_new",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_MUTEX_NEW",
-        args: &[AbiType::I64],
-        returns: AbiType::Handle,
-        doc: "Aloca um Mutex protegendo um valor i64 inicializado com `initial`.",
-        ts_signature: "mutex_new(initial: number): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "mutex_lock",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_MUTEX_LOCK",
-        args: &[AbiType::U64],
-        returns: AbiType::I64,
-        doc: "Bloqueia ate adquirir o Mutex e retorna o valor interno protegido. 0 se handle invalido.",
-        ts_signature: "mutex_lock(mutex: number): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "mutex_try_lock",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_MUTEX_TRY_LOCK",
-        args: &[AbiType::U64],
-        returns: AbiType::I64,
-        doc: "Tenta adquirir o Mutex sem bloquear. Retorna o valor interno em caso de sucesso, 0 se ja estava lockado ou handle invalido.",
-        ts_signature: "mutex_try_lock(mutex: number): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "mutex_set",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_MUTEX_SET",
-        args: &[AbiType::U64, AbiType::I64],
-        returns: AbiType::Void,
-        doc: "Escreve `value` no Mutex. Caller deve ter chamado lock/try_lock antes (responsabilidade do caller).",
-        ts_signature: "mutex_set(mutex: number, value: number): void",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "mutex_unlock",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_MUTEX_UNLOCK",
-        args: &[AbiType::U64],
-        returns: AbiType::Void,
-        doc: "Libera o Mutex previamente adquirido por lock/try_lock. No-op se nao havia guard ativo.",
-        ts_signature: "mutex_unlock(mutex: number): void",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "mutex_free",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_MUTEX_FREE",
-        args: &[AbiType::U64],
-        returns: AbiType::Void,
-        doc: "Libera o Mutex e seu slot na HandleTable.",
-        ts_signature: "mutex_free(mutex: number): void",
-        intrinsic: None,
-        pure: false,
-    },
-    // ── RwLock ─────────────────────────────────────────────────────
-    NamespaceMember {
-        name: "rwlock_new",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_RWLOCK_NEW",
-        args: &[AbiType::I64],
-        returns: AbiType::Handle,
-        doc: "Aloca um RwLock protegendo um valor i64 inicializado com `initial`.",
-        ts_signature: "rwlock_new(initial: number): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "rwlock_read",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_RWLOCK_READ",
-        args: &[AbiType::U64],
-        returns: AbiType::Handle,
-        doc: "Adquire um read guard (compartilhado) e retorna um handle de guard. Liberar via rwlock_unlock(guard). 0 se handle invalido.",
-        ts_signature: "rwlock_read(rwlock: number): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "rwlock_write",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_RWLOCK_WRITE",
-        args: &[AbiType::U64],
-        returns: AbiType::Handle,
-        doc: "Adquire um write guard (exclusivo) e retorna um handle de guard. Liberar via rwlock_unlock(guard). 0 se handle invalido.",
-        ts_signature: "rwlock_write(rwlock: number): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "rwlock_unlock",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_RWLOCK_UNLOCK",
-        args: &[AbiType::U64],
-        returns: AbiType::Void,
-        doc: "Libera um guard previamente adquirido via rwlock_read/rwlock_write.",
-        ts_signature: "rwlock_unlock(guard: number): void",
-        intrinsic: None,
-        pure: false,
-    },
-    // ── OnceLock ───────────────────────────────────────────────────
-    NamespaceMember {
-        name: "once_new",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_ONCE_NEW",
-        args: &[],
-        returns: AbiType::Handle,
-        doc: "Aloca um OnceLock e retorna o handle.",
-        ts_signature: "once_new(): number",
-        intrinsic: None,
-        pure: false,
-    },
-    NamespaceMember {
-        name: "once_call",
-        kind: MemberKind::Function,
-        symbol: "__RTS_FN_NS_SYNC_ONCE_CALL",
-        args: &[AbiType::U64, AbiType::I64],
-        returns: AbiType::Void,
-        doc: "Executa `fn_ptr` (ponteiro para `extern \"C\" fn()`) exatamente uma vez por OnceLock. Chamadas subsequentes sao no-op.",
-        ts_signature: "once_call(once: number, fn_ptr: number): void",
-        intrinsic: None,
-        pure: false,
-    },
-];
-
-pub const SPEC: NamespaceSpec = NamespaceSpec {
-    name: "sync",
-    doc: "Primitivas de sincronizacao (Mutex, RwLock, OnceLock) baseadas em std::sync.",
-    members: MEMBERS,
-};
